requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
coverage==4.5.1
coveralls==1.5.0
docopt==0.6.2
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.13.1
Flask-RESTful==0.3.6
Flask-Testing==0.7.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
Flask-HTTPAuth==3.2.4
itsdangerous==0.24
email-validator==1.0.3
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==2.15.6
marshmallow-jsonapi==0.20.3
mccabe==0.6.1
more-itertools==4.3.0
passlib==1.7.1
pep8==1.7.1
pluggy==0.7.1
postgres==2.2.2
psycopg2-binary==2.7.5
py==1.6.0
pycodestyle==2.4.0
PyJWT==1.4.2
pylint==2.1.1
pytest==3.8.0
pytest-cov==2.6.0
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
